+++ R/plot_rich_norm.R
@@ -23,14 +23,17 @@
 
 plot_rich_norm <- function(rich.dt, facet.by=c('density', 'threshold'),
                            densities, alpha=0.05, g=NULL) {
-  p.fdr <- yloc <- Group <- phi <- xstart <- xend <- NULL
+  p.fdr <- yloc <- Group <- phi <- xstart <- xend <- Study.ID <- NULL
 
+  facet.by <- match.arg(facet.by)
   subDT <- rich.dt[eval(parse(text=facet.by)) %in% densities]
   subDT[, star := ifelse(p.fdr < alpha, '*', '')]
-  subDT[, yloc := lapply(.SD, function(x)
-                         min(x, na.rm=T) - 0.05 * diff(range(x, na.rm=T))),
-        by=density, .SDcols='phi']
-  subDT[Group == levels(subDT$Group)[2], yloc := yloc - 0.005]
+  subDT[, yloc := min(phi, na.rm=T) - 0.05 * diff(range(phi, na.rm=T)), by=facet.by]
+  if (nlevels(subDT$Group) > 1) {
+    for (i in 2:nlevels(subDT$Group)) {
+      subDT[Group == levels(subDT$Group)[i], yloc := yloc - i * 0.05 * diff(range(phi, na.rm=T))]
+    }
+  }
   setkeyv(subDT, c(facet.by, 'Group'))
 
   if (!is.null(g)) {
@@ -53,9 +56,16 @@ plot_rich_norm <- function(rich.dt, facet.by=c('density', 'threshold'),
                         Group=rep(subDT[, unique(Group)],
                                   each=length(densities)))
   }
-  setkey(rects, density, Group)
+  setnames(rects, 'density', facet.by)
+  setkeyv(rects, c(facet.by, 'Group'))
 
-  p <- ggplot(data=subDT[rects], aes(x=k, y=phi)) +
+  if ('Study.ID' %in% names(subDT)) {
+    rects <- subDT[rects]
+    p <- ggplot(data=rects, aes(x=k, y=phi, group=Study.ID))
+  } else {
+    p <- ggplot(data=subDT[rects], aes(x=k, y=phi))
+  }
+  p <- p +
     geom_line(aes(col=Group)) +
     geom_hline(yintercept=1, size=0.5, lty=2) +
     geom_text(aes(y=yloc, col=Group, label=star), size=5, show.legend=F) +
@@ -64,8 +74,7 @@ plot_rich_norm <- function(rich.dt, facet.by=c('density', 'threshold'),
               alpha=0.08, fill='red') +
     facet_wrap(as.formula(paste('~', facet.by)), scales='free') +
     xlab('Degree (k)') + ylab(expression(phi[norm])) +
-    theme(legend.position=c(0, 1), legend.justification=c(0, 1),
-          legend.background=element_rect(size=0.5))
+    theme(legend.position='bottom')
 
   return(p)
 }